Adds local
Local is a containerized version of the setup i had on my local
development machine. It mounts webcontent and certificates (self-signed,
trusted by my local browser) not so useful but good enough to keep
iterating on a more production like setup. What I would need to do:
1. Publish the images so I can get at them easily
2. Use let's encrypt or something
3. Buy metal on digital ocean

Perhaps a next step is to do a setup on a linux-box on the local
network.